+++ 08.ls_object/regular_expression.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: false
+
+class RegularExpression
+  def self.push_file_that_matches_regular_expression(file_list, regex)
+    file_list.select { |file| file.match?(regex) }
+  end
+end
